ab_contracts_executor

Struct NativeExecutor

pub struct NativeExecutor { /* private fields */ }

Implementations

impl NativeExecutor

pub fn new_storage(&self) -> Result<Storage, ContractError>

Create a new storage instance with system contracts already deployed

pub fn builder(shard_index: ShardIndex) -> NativeExecutorBuilder

Builder of native executor for specified shard index

pub fn transaction_verify( &self, transaction: Transaction<'_>, storage: &Storage, ) -> Result<(), ContractError>

Verify the provided transaction.

Self::transaction_execute() can be used for transaction execution if needed. Self::transaction_verify_execute() can be used to verify and execute a transaction with a single call.

pub fn transaction_execute( &self, transaction: Transaction<'_>, storage: &mut Storage, ) -> Result<(), ContractError>

Execute the previously verified transaction.

Self::transaction_verify() must be used for verification. Self::transaction_verify_execute() can be used to verify and execute a transaction with a single call.

pub fn transaction_verify_execute( &self, transaction: Transaction<'_>, storage: &mut Storage, ) -> Result<(), ContractError>

Verify and execute provided transaction.

A shortcut for Self::transaction_verify() + Self::transaction_execute().

pub fn transaction_emulate<Calls, T>( &self, contract: Address, storage: &mut Storage, calls: Calls, ) -> T
where Calls: FnOnce(&mut Env<'_>) -> T,

Emulate a transaction submitted by contract with method calls happening inside calls without going through TxHandler.

NOTE: This is primarily useful for testing environment, usually changes are done in the transaction execution using Self::transaction_execute().

pub fn env_ro<'a>(&'a self, storage: &'a Storage) -> Env<'a>

Get a read-only Env instance for calling #[view] methods on it directly.

For stateful methods, execute a transaction using Self::transaction_execute() or emulate one with Self::transaction_emulate().